refactor(payroll): PayrollCalculator 1513→630 lignes — extraction des valeurs de bulletin dans PaySlipValueCalculator (suite #5591, slice 2)
Suite du découpage du god-object : le concern « assemblage des valeurs du
bulletin » (computeSlipValues + ses 16 helpers : computeWorkedDays,
computeProratedBase, computeOvertimePay, computeSickLeaveAllowance,
computeLeaveIndemnity, referenceGross12Months, accruedLeaveDays,
computeNetBreakdown, slabTaxBreakdown, proratas…) est extrait dans
PaySlipValueCalculator, comportement STRICTEMENT identique (copie stricte).

- PayrollCalculator : l'orchestration (calculateRun, exécution, régularisation,
  calculateSlip, computeFinalSettlement) + 7 façades publiques (API stable des
  simulateurs/tests golden) ; les constantes STANDARD_WORKING_DAYS /
  MONTHLY_HOURS restent exposées (alias vers le service).
- PaySlipValueCalculator : calcul pur, résolveur de règles pays partagé,
  agrégateur d'entrées de travail partagé (même instance).
- Filet : 476 tests verts (448 golden multi-pays) + suite paie/simulation
  complète : 0 régression vs main (10 échecs pré-existants identiques).
